Time: 1 ms (98.96%), Space: 42.5 MB (78.87%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -1,12 +1,11 @@
 class Solution {
     public int mySqrt(int x) {
         int low = 1, high = x;
-
         while(low <= high){
             int mid = low + (high - low) / 2;
-            long val = (long) mid * (long) mid;
-
-            if(val <= (long) x){
+            long val = (long)mid * (long)mid;
+            
+            if(val <= x){
                 low = mid + 1;
             }else{
                 high = mid - 1;
